Refactored the controllers tests to move post creation functions to the base
